아이템 전달 기능 추가
ㅁㄴㅇㄹ

+++ src/items/items.commands.ts
@@ -15,7 +15,11 @@ import {
   ModalBuilder,
   TextInputBuilder,
 } from 'discord.js';
-import { ItemsService, ItemTransactionResult } from './items.service';
+import {
+  ItemsService,
+  ItemTransactionResult,
+  GiveResult,
+} from './items.service';
 import { ItemsTransactionsInfoDTO } from './ItemsTransactionsInfo-dto';
 import { ALLOWED } from 'src/gold/gold.commands';
 import { GoldService } from 'src/gold/gold.service';
@@ -27,6 +31,7 @@ import { CharacterNameDTO } from 'src/gold/CharacterName-dto';
 import { CharacterGoldDTO } from 'src/gold/CharacterGold-dto';
 import { ItemsPageDTO } from './ItemsPage-dto';
 import { ItemsAliasDTO } from './ItemAlias-dto';
+import { ItemsTradeInfoDTO } from './ItemsTradeInfo-dto';
 
 @Injectable()
 export class ItemsCommands {
@@ -378,7 +383,7 @@ export class ItemsCommands {
       const result =
         await this.itemsService.TryUseItemInventory(itemsUseInfoDTO);
       if (result.scenario === 0) {
-        console.log(`시나리오 0번`);
+        // console.log(`시나리오 0번`);
         return interaction.reply({
           content:
             this.goldService.StringFormatter(
@@ -408,7 +413,7 @@ export class ItemsCommands {
             ),
         });
       } else if (result.scenario === 1) {
-        console.log(`시나리오 1번`);
+        // console.log(`시나리오 1번`);
         return interaction.reply({
           content:
             `-# ${result.itemName}, ${result.owner}, ${result.amount}, ${result.quality}, ${result.unit}\n` +
@@ -801,6 +806,20 @@ export class ItemsCommands {
     guilds: ['1284642997375336592', '1273347630767804539'],
   })
   public async onCleanDummy(@Context() [interaction]: SlashCommandContext) {
+    if (!ALLOWED.has(interaction.user.id)) {
+      await interaction.reply({
+        content: this.goldService.StringFormatter(
+          `🚫 커맨드를 사용할 권한이 없습니다.\n 관리자가 아닐 경우, 데이터 읽기만 가능합니다.\n` +
+            ` 사용자 ID: ${interaction.user.id}`,
+          TextColor.BOLD_RED,
+          true,
+          true,
+        ),
+        flags: 'Ephemeral',
+      });
+      return;
+    }
+
     try {
       const result = await this.itemsService.purgeDummyInventory();
       return interaction.reply({
@@ -811,6 +830,162 @@ export class ItemsCommands {
           true,
         ),
       });
+    } catch (err: any) {
+      return interaction.reply({
+        content: this.goldService.StringFormatter(
+          `🚫 에러 발생: ` + err.message,
+          TextColor.BOLD_RED,
+          true,
+          true,
+        ),
+      });
+    }
+  }
+
+  @SlashCommand({
+    name: 'item-give',
+    description: `transferring item from a character to character`,
+    guilds: ['1284642997375336592', '1273347630767804539'],
+  })
+  public async onGiveItem(
+    @Context() [interaction]: SlashCommandContext,
+    @Options() itemsTradeInfoDTO: ItemsTradeInfoDTO,
+  ) {
+    if (!ALLOWED.has(interaction.user.id)) {
+      await interaction.reply({
+        content: this.goldService.StringFormatter(
+          `🚫 커맨드를 사용할 권한이 없습니다.\n 관리자가 아닐 경우, 데이터 읽기만 가능합니다.\n` +
+            ` 사용자 ID: ${interaction.user.id}`,
+          TextColor.BOLD_RED,
+          true,
+          true,
+        ),
+        flags: 'Ephemeral',
+      });
+      return;
+    }
+
+    try {
+      const result = new GiveResult();
+      await this.itemsService.TryGiveItem(itemsTradeInfoDTO, result);
+
+      if (result.scenario === 0) {
+        // 정상 입력
+        // console.log(`시나리오 0`);
+        return interaction.reply({
+          content:
+            this.goldService.StringFormatter(
+              `📦 [아이템 전달 이벤트 발생 알림]`,
+              TextColor.BOLD_WHITE,
+              true,
+              false,
+            ) +
+            '\n' +
+            this.goldService.StringFormatter(
+              `「${result.from}」, 「${result.to}」에게 `,
+              TextColor.BOLD_WHITE,
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `[${result.itemName}]`,
+              this.ColorParser(result.quality),
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `를 ${result.moved}${result.unit}만큼 전달하였다.\n`,
+              TextColor.BOLD_WHITE,
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `「${result.from}」, [${result.itemName}]의 남은 수량: ${result.fromRemaining}${result.unit}\n`,
+              TextColor.BOLD_GRAY,
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `「${result.to}」, [${result.itemName}]의 남은 수량: ${result.toTotal}${result.unit}`,
+              TextColor.BOLD_GRAY,
+              false,
+              true,
+            ),
+        });
+      } else if (result.scenario === 1) {
+        return interaction.reply({
+          content:
+            `-# ${result.from}, ${result.to}, ${result.itemName}, ${result.moved}, ${result.quality}, ${result.unit}\n` +
+            this.goldService.StringFormatter(
+              `[${result.itemName}]`,
+              this.ColorParser(result.quality),
+              true,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `(을)를 의도하신 것 같아요. 맞나요?`,
+              TextColor.BOLD_WHITE,
+              false,
+              true,
+            ),
+          components: [
+            new ActionRowBuilder<ButtonBuilder>().addComponents(
+              new ButtonBuilder()
+                .setCustomId('YES_BUTTON_ITEM_TRANSFER')
+                .setLabel('네')
+                .setStyle(ButtonStyle.Primary),
+            ),
+            new ActionRowBuilder<ButtonBuilder>().addComponents(
+              new ButtonBuilder()
+                .setCustomId('NO_BUTTON')
+                .setLabel('아니오')
+                .setStyle(ButtonStyle.Primary),
+            ),
+          ],
+        });
+      } else if (result.scenario === 3) {
+        // 앨리어싱
+        return interaction.reply({
+          content:
+            this.goldService.StringFormatter(
+              `📦 [아이템 전달 이벤트 발생 알림]`,
+              TextColor.BOLD_WHITE,
+              true,
+              false,
+            ) +
+            '\n' +
+            this.goldService.StringFormatter(
+              `「${result.from}」, 「${result.to}」에게 `,
+              TextColor.BOLD_WHITE,
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `[${itemsTradeInfoDTO.itemName} (${result.itemName})]`,
+              this.ColorParser(result.quality),
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `를 ${result.moved}${result.unit}만큼 전달하였다.\n`,
+              TextColor.BOLD_WHITE,
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `「${result.from}」, [${result.itemName}]의 남은 수량: ${result.fromRemaining}${result.unit}\n`,
+              TextColor.BOLD_GRAY,
+              false,
+              false,
+            ) +
+            this.goldService.StringFormatter(
+              `「${result.to}」, [${result.itemName}]의 남은 수량: ${result.toTotal}${result.unit}`,
+              TextColor.BOLD_GRAY,
+              false,
+              true,
+            ),
+        });
+      }
     } catch (err: any) {
       return interaction.reply({
         content: this.goldService.StringFormatter(
